+++ CHANGELOG.md
@@ -9,6 +9,7 @@ HumHub Changelog
 - Fix #7471: Fix advanced searching by space filter
 - Fix #7472: Fix missing fields when creating a new user from admin
 - Fix #7477: Refactor Registration Form Options 
+- Enh #7455: Add Share Intend feature for the Mobile app
 
 1.17.1 (March 6, 2025)
 ----------------------